MoneyThing/requirements.txt
pandas>=2.3.1
streamlit>=1.48.1
streamlit-aggrid>=1.1.8.post1
pdfplumber>=0.11.7
numpy>=2.2.6
bcrypt>=4.2.0
pytesseract~=0.3.13
pillow~=11.3.0
plotly>=5.18.0